requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25

# Redis
redis==5.0.1

# OCR
easyocr==1.7.0
Pillow==10.2.0
opencv-python-headless==4.9.0.80

# AI
openai==1.12.0

# WhatsApp (Twilio)
twilio==8.11.1

# Security
cryptography==42.0.2

# Utilities
python-dotenv==1.0.0
requests==2.31.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
